subindo wb

+++ src/main/java/com/tupa/restaurante/entidades/Produto.java
@@ -5,6 +5,8 @@ import lombok.Setter;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Document(collection = "Produtos")
@@ -12,7 +14,7 @@ public class Produto {
     @Id
     private String id;
     private String nome;
-    private double preco;
+    private BigDecimal preco;
     private String tipo; // Tipo do produto: CHURRASCO, BEBIDA, COZINHA, etc.
     private String imagem; // Campo para armazenar a imagem em Base64
 }
